requirements.txt
llama-cpp-python==0.2.70
langchain==0.1.20
langchain-community==0.0.38
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic<2
huggingface-hub==0.25.2
pandas